Ooops, undo a previous privatization
