+++ opencog/atoms/execution/ExecSCM.cc
@@ -15,30 +15,7 @@
 #include <opencog/atoms/reduct/FoldLink.h>
 #include <opencog/guile/SchemeModule.h>
 
-#ifndef _OPENCOG_EXEC_SCM_H
-#define _OPENCOG_EXEC_SCM_H
-
-#include <opencog/guile/SchemeModule.h>
-
-namespace opencog {
-
-class ExecSCM : public ModuleWrap
-{
-	protected:
-		virtual void init(void);
-		static std::vector<FunctionWrap*> _binders;
-	public:
-		ExecSCM(void);
-		~ExecSCM();
-};
-
-}
-
-extern "C" {
-void opencog_exec_init(void);
-};
-
-#endif // _OPENCOG_EXEC_SCM_H
+#include "ExecSCM.h"
 
 // ========================================================
